Extract the logic for calculating metrics into a file
